requirements.txt
pybit
pandas
pandas-ta
python-dotenv
PyYAML
python-telegram-bot
backtesting
SQLAlchemy
alembic
psycopg2-binary
redis
tenacity